backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
python-dotenv==1.0.1
google-generativeai==0.7.2
python-multipart==0.0.9
pydantic==2.8.2
aiofiles==23.2.1
